requirements.txt
prometheus_client
Flask
pyyaml
fire